# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hubuum REST API",
    "description": "OpenAPI documentation for the Hubuum REST service.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment base URL"
    }
  ],
  "paths": {
...
    "/api/v1/namespaces": {
      "get": {
        "tags": [
          "namespaces"
        ],
        "summary": "Get Api V1 Namespaces",
        "description": "Auto-generated documentation for GET /api/v1/namespaces. Supports cursor pagination through the `limit`, `sort`, and `cursor` query parameters. The next page cursor is returned in the `X-Next-Cursor` response header.",
        "operationId": "getApiV1Namespaces",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 100. Maximum is 250.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields. Cursor pagination uses the requested sort order and appends a stable tie-breaker automatically.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in the X-Next-Cursor response header from a previous page. Supply it unchanged to fetch the next page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Namespaces matching optional query filters The response body contains the current page items as a JSON array. Use the `X-Next-Cursor` header, when present, to request the next page.",
            "headers": {
              "X-Next-Cursor": {
                "schema": {
                  "type": "string"
                },
                "description": "Opaque cursor for the next page. This header is omitted when there are no more results."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Namespace"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "namespaces"
        ],
        "summary": "Post Api V1 Namespaces",
        "description": "Auto-generated documentation for POST /api/v1/namespaces.",
        "operationId": "postApiV1Namespaces",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewNamespaceWithAssignee"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Namespace created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Namespace"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/namespaces/{namespace_id}/has_permissions/{permission}": {
      "get": {
        "tags": [
          "namespaces"
        ],
        "summary": "List all groups that have any permissions on a namespace",
        "description": "Auto-generated documentation for GET /api/v1/namespaces/{namespace_id}/has_permissions/{permission}. Supports cursor pagination through the `limit`, `sort`, and `cursor` query parameters. The next page cursor is returned in the `X-Next-Cursor` response header.",
        "operationId": "getApiV1NamespacesByNamespaceIdHasPermissionsByPermission",
        "parameters": [
          {
            "name": "namespace_id",
            "in": "path",
            "description": "Namespace ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "permission",
            "in": "path",
            "description": "Permission value",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Permissions"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 100. Maximum is 250.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields. Cursor pagination uses the requested sort order and appends a stable tie-breaker automatically.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in the X-Next-Cursor response header from a previous page. Supply it unchanged to fetch the next page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Groups with permission The response body contains the current page items as a JSON array. Use the `X-Next-Cursor` header, when present, to request the next page.",
            "headers": {
              "X-Next-Cursor": {
                "schema": {
                  "type": "string"
                },
                "description": "Opaque cursor for the next page. This header is omitted when there are no more results."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Namespace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "error": "Unauthorized",
          "message": "Authentication failure"
        }
      },
...
      "Group": {
        "type": "object",
        "required": [
          "id",
          "groupname",
          "description",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "groupname": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Namespace": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NewNamespaceWithAssignee": {
        "type": "object",
        "description": "A new namespace, with an assignee. Used for creating new namespace entries\ninto the database and assign all permissions to the group given as group_id.\n\nThis wraps the NewNamespace struct and uses the group_id to grant all permissions\nto the group in a single transaction.",
        "required": [
          "name",
          "description",
          "group_id"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "group_id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        },
        "example": {
          "description": "Shared assets and metadata",
          "group_id": 1,
          "name": "global-assets"
        }
      },
...
      "Permissions": {
        "type": "string",
        "enum": [
          "ReadCollection",
          "UpdateCollection",
          "DeleteCollection",
          "DelegateCollection",
          "CreateClass",
          "ReadClass",
          "UpdateClass",
          "DeleteClass",
          "CreateObject",
          "ReadObject",
          "UpdateObject",
          "DeleteObject",
          "CreateClassRelation",
          "ReadClassRelation",
          "UpdateClassRelation",
          "DeleteClassRelation",
          "CreateObjectRelation",
          "ReadObjectRelation",
          "UpdateObjectRelation",
          "DeleteObjectRelation"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}